requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
openpyxl==3.1.5
psycopg2==2.9.9
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==1.0.1
Werkzeug==3.0.4